Window + Event/IO Loop (With Logging)
That took longer than expected.  It should be a good platform, though.  It's
not trivial to integrate networking into the pygame event loop; hopefully
this will work reasonably.  I decided to continuously process the network
because I think that pygame's event buffer can grow to whatever size it needs.
I'm enforcing 30FPS mostly to ensure that everyone in a network game runs at
the same speed.  Latency will still be an issue, though; I'll cross that
bridge when I come to it.  I've wasted enough time already on this network
stack, when I might not even have time to make it actually work.