Beginnings of an event system; attract mode
I've established a grid to play on, and can paint the individual cells of that
grid whatever color I want.  Also, it quits cleanly now and can translate mouse
coordinates into grid coordinates.  I don't know if I actually want to use the
mouse in the game, but it was a good test to make sure I understood the
coordinate system.