Command-line options
The networking options don't do anything yet, but you can now control the
verbosity of the log messages, as well as where the log goes.  I should
probably only log the random number seed, and then not bother with the
frame-by-frame report of which cells get filled -- that puts an awful lot
of output in the debug log.  In theory, though, the debug log should have
all of the information necessary to replay the entire session in it.  At
some point I should write a method to read one of the debug logs and replay
the game.  I probably won't worry about it too much, because this isn't
really part of the final game; it's just an interesting way for me to test
the underlying engine.  That seems to be going rather slowly, though, so
it's entirely possible that it's the only thing I'll be able to finish by
the deadline.

In any case, my next chore is to get two copies of this funny box-popping
thing running in sync.