Networking!
I really spent a lot of time on this; I hope it's as robust as I think it is.
Once per frame, each side sends a ping packet that includes both the local
frame number and the most-recently-seen remote frame number.  After 30 of
these have been exchanged (1sec), the server decides on a correlation between
the two sequences of frame numbers, and both sides use that to figure out
when things should happen.  All events are fired at the same time on both
computers, in a deterministic order.  To do this, locally-generated events
are delayed 5 frames before they take effect.  This scheme should be able to
accomodate up to 10 frames of round-trip ping time (330 ms).  This can be
increased if the ping times over the general internet are higher than this.