Factor out game logic into a separate file
Just a cleanup release; I think that all of the event handling and networking
code is basically complete.  I moved all of the code that actually does any
game logic into world.py, and set up the main file to call functions in it.

As you'd expect, if render_frame tries to change the state of the world, then
things will break.  Note that this includes generating random numbers from the
defauld RNG.  I should probably use a separate one for the shared state anyway.
Unfortunately, they won't break until there's a dropped frame, so use caution.

Also, the raw event handlers should be pure input handlers.  They shouldn't do
anything that might modify the shared state.  The only things that they should
be doing is posting generic events which will be executed simultaneously on
both sides of the network.  The only exception to this is state that is needed
purely for UI purposes, so that higher-level events can be generated and
forwarded.